# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services": {
      "get": {
        "operationId": "ServicesController_adminCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Полный каталог для админки",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "ServicesController_createService",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать услугу",
        "tags": [
          "services"
        ]
      }
    },
    "/services/groups": {
      "post": {
        "operationId": "ServicesController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceGroupBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceGroupViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать группу",
        "tags": [
          "services"
        ]
      }
    },
    "/services/groups/order": {
      "put": {
        "operationId": "ServicesController_reorderGroups",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Порядок групп",
        "tags": [
          "services"
        ]
      }
    },
    "/services/groups/{id}": {
      "patch": {
        "operationId": "ServicesController_updateGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceGroupBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceGroupViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Изменить группу",
        "tags": [
          "services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_deleteGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить группу с услугами",
        "tags": [
          "services"
        ]
      }
    },
    "/services/order": {
      "put": {
        "operationId": "ServicesController_reorderServices",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicesOrderBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Порядок услуг в группе",
        "tags": [
          "services"
        ]
      }
    },
    "/services/{id}": {
      "patch": {
        "operationId": "ServicesController_updateService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Изменить услугу",
        "tags": [
          "services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_deleteService",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить услугу",
        "tags": [
          "services"
        ]
      }
    },
    "/services/showcase": {
      "put": {
        "operationId": "ServicesController_updateShowcase",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShowcaseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShowcaseSlotViewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Задать витрину (ровно 8)",
        "tags": [
          "services"
        ]
      }
    },
...
  },
  "info": {
    "title": "STO API",
    "description": "API автосервиса: публичный сайт и админка",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "локальная разработка"
    }
  ],
  "components": {
    "schemas": {
...
      "ServiceViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "zamena-masla"
          },
          "groupId": {
            "type": "string",
            "example": "shinomontazh"
          },
          "title": {
            "type": "string",
            "example": "Замена масла"
          },
          "description": {
            "type": "string",
            "example": "Синтетика"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000
          },
          "priceFrom": {
            "type": "number",
            "example": 60
          },
          "priceTo": {
            "type": "object",
            "nullable": true,
            "example": 120
          },
          "unit": {
            "type": "string",
            "nullable": true,
            "enum": [
              "wheel",
              "pcs",
              "season"
            ]
          },
          "photo": {
            "type": "object",
            "nullable": true,
            "example": "/uploads/oil.webp"
          },
          "position": {
            "type": "number",
            "example": 0
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "groupId",
          "title",
          "description",
          "details",
          "priceFrom",
          "priceTo",
          "unit",
          "photo",
          "position",
          "isActive"
        ]
      },
      "ServiceGroupViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "shinomontazh"
          },
          "title": {
            "type": "string",
            "example": "Шиномонтаж"
          },
          "icon": {
            "type": "string",
            "nullable": true,
            "enum": [
              "scan",
              "wrench",
              "disc",
              "cog",
              "gauge",
              "circle",
              "snowflake"
            ]
          },
          "photo": {
            "type": "object",
            "nullable": true,
            "example": "/uploads/tyres.webp"
          },
          "position": {
            "type": "number",
            "example": 0
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceViewDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "icon",
          "photo",
          "position",
          "isActive",
          "services"
        ]
      },
      "ShowcaseSlotViewDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "example": "zamena-masla"
          },
          "position": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "serviceId",
          "position"
        ]
      },
      "CatalogResponseDto": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceGroupViewDto"
            }
          },
          "showcase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShowcaseSlotViewDto"
            }
          }
        },
        "required": [
          "groups",
          "showcase"
        ]
      },
      "CreateServiceGroupBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Шиномонтаж"
          },
          "icon": {
            "type": "string",
            "enum": [
              "scan",
              "wrench",
              "disc",
              "cog",
              "gauge",
              "circle",
              "snowflake"
            ],
            "nullable": true
          },
          "photo": {
            "type": "object",
            "nullable": true,
            "example": "/uploads/tyres.webp"
          }
        },
        "required": [
          "title"
        ]
      },
      "OrderBody": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "UpdateServiceGroupBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Шиномонтаж"
          },
          "icon": {
            "type": "string",
            "enum": [
              "scan",
              "wrench",
              "disc",
              "cog",
              "gauge",
              "circle",
              "snowflake"
            ],
            "nullable": true
          },
          "photo": {
            "type": "object",
            "nullable": true,
            "example": "/uploads/tyres.webp"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "title"
        ]
      },
      "CreateServiceBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Замена масла"
          },
          "description": {
            "type": "string",
            "example": "Синтетика"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000,
            "example": "В стоимость входит масло, фильтр и работа мастера…"
          },
          "priceFrom": {
            "type": "number",
            "example": 60
          },
          "priceTo": {
            "type": "object",
            "nullable": true,
            "example": 120
          },
          "unit": {
            "type": "string",
            "enum": [
              "wheel",
              "pcs",
              "season"
            ],
            "nullable": true
          },
          "photo": {
            "type": "object",
            "nullable": true,
            "example": "/uploads/oil.webp"
          },
          "groupId": {
            "type": "string",
            "example": "shinomontazh"
          }
        },
        "required": [
          "title",
          "priceFrom",
          "groupId"
        ]
      },
      "ServicesOrderBody": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "groupId": {
            "type": "string",
            "example": "shinomontazh"
          }
        },
        "required": [
          "ids",
          "groupId"
        ]
      },
      "UpdateServiceBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Замена масла"
          },
          "description": {
            "type": "string",
            "example": "Синтетика"
          },
          "details": {
            "type": "object",
            "nullable": true,
            "maxLength": 2000,
            "example": "В стоимость входит масло, фильтр и работа мастера…"
          },
          "priceFrom": {
            "type": "number",
            "example": 60
          },
          "priceTo": {
            "type": "object",
            "nullable": true,
            "example": 120
          },
          "unit": {
            "type": "string",
            "enum": [
              "wheel",
              "pcs",
              "season"
            ],
            "nullable": true
          },
          "photo": {
            "type": "object",
            "nullable": true,
            "example": "/uploads/oil.webp"
          },
          "groupId": {
            "type": "string",
            "example": "shinomontazh"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "title",
          "priceFrom",
          "groupId"
        ]
      },
      "ShowcaseBody": {
        "type": "object",
        "properties": {
          "serviceIds": {
            "minItems": 8,
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "serviceIds"
        ]
      },
...
    }
  }
}